add parser policy
To support for future file format changes, we add an option to
allow parser try to parse for files with unknown signature.

The default policy is 'strict' which raises exception for
unknown signature.  When you set 'loose' policy, the parser
try to parse for unknown signature.

Note that loose policy may cause unexpected output.